ai_librarian_monorepo/ai_librarian_core/pyproject.toml
[project]
name = "ai-librarian-core"
version = "0.1.1"
description = "A wrapper package for LangChain and LangGraph."
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "duckduckgo-search>=8.0.2",
    "google-api-python-client>=2.177.0",
    "httpx>=0.28.1",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.61",
    "langchain-google-community>=2.0.7",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "langsmith>=0.3.42",
    "lxml>=6.0.0",
    "playwright>=1.52.0",
    "pydantic>=2.11.5",
    "pyowm>=3.3.0",
    "setuptools>=80.9.0", # PyOWM requires setuptools
    "wikipedia>=1.4.0",
    "youtube-search>=2.1.2",
    "rich>=14.1.0",
    "langchain-anthropic>=0.3.18",
    "langchain-google-genai>=2.1.8",
    "langchain-groq>=0.3.6",
]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
